--- index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>WebGL2</title>
-        <link rel="stylesheet" href="./css/style.css">
-        <link rel="shortcut icon" href="./img/cowicon.png">
-    </head>
-    <body>
-        <canvas id="canvas"></canvas>
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/gl-matrix/2.8.1/gl-matrix-min.js"></script>
-        <script src="./js/gl-utils.js"></script>
-        <script src="./js/main.js"></script>
-    </body>
-</html>
