+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "webgl-2-background",
+  "description": "Create a background for a website using webgl2",
+  "version": "1.0.0",
+  "author": "",
+  "license": "ISC",
+  
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  
+  "dependencies": {
+    "gl-matrix": "^3.4.3"
+  },
+
+  "devDependencies": {
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
